Renaming & Summary improvements
