+++ WalletWasabi.Fluent/App.axaml.cs
@@ -7,6 +7,7 @@ using Avalonia.Controls.ApplicationLifetimes;
 using Avalonia.Markup.Xaml;
 using NBitcoin;
 using ReactiveUI;
+using WalletWasabi.Announcements;
 using WalletWasabi.Fluent.Models;
 using WalletWasabi.Fluent.Models.ClientConfig;
 using WalletWasabi.Fluent.Models.FileSystem;
@@ -134,7 +135,7 @@ public class App : Application
 		var applicationSettings = CreateApplicationSettings();
 		var torStatusChecker = new TorStatusCheckerModel();
 
-		var releaseHighlights = new ReleaseHighlights.ReleaseHighlights();
+		var releaseHighlights = new ReleaseHighlights();
 
 		// This class (App) represents the actual Avalonia Application and it's sole presence means we're in the actual runtime context (as opposed to unit tests)
 		// Once all ViewModels have been refactored to receive UiContext as a constructor parameter, this static singleton property can be removed.
