+++ WalletWasabi/Announcements/ReleaseHighlights.md
@@ -1,20 +1,20 @@
 ## Release Highlights
-🪲   Bug Fixes<br/>
-💫   Settings Layout & UI Enhancements<br/>
-⤴️   Tor Update: Upgraded to v13.5.9<br/>
-🚫   Buy Anything Button Disabled<br/>
+#### 🪲 Bug Fixes<br/>
+#### 💫 Settings Layout & UI Enhancements<br/>
+#### 📈 Tor Update: Upgraded to v13.5.9<br/>
+#### 🚫 Buy Anything Button Disabled<br/>
 
 ## Release Summary
 Wasabi Wallet v2.3.1 is a stability-focused release packed with critical bug fixes and preparatory changes for upcoming major features.
 
-### 🪲   Notable Bug Fixes
+### 🪲 Notable Bug Fixes
 - Fixed an issue where the transaction broadcaster displayed an error, while the transaction was successfully broadcast
 - Resolved a problem with DNS endpoints for remote Bitcoin nodes
 - Improved the amount decomposer to prevent privacy leaks when using the *payment-in-round* feature
 - Fixed sorting tables by amount
 - Sometimes small rounds were not recognized by the wallet, leading to improper anon score computation
 
-### 💫   Settings & UI Enhancements
+### 💫 Settings & UI Enhancements
 The Settings layout has been refined, optimizing space for a more streamlined user experience.
 
 Additionally, some new UI features have been implemented:
@@ -22,7 +22,7 @@ Additionally, some new UI features have been implemented:
 - Added functionality to copy addresses directly from the Coin List view
 - Consistent formatting of Fee Rate to 2 decimals without trailing 0
 
-### 🚫   Buy Anything Button Disabled
+### 🚫 Buy Anything Button Disabled
 This feature, launched nearly a year ago, allowed users to access the premium ShopinBit concierge service directly through the wallet.
 
 This version disables it, for the following reasons:
